+++ cml-compiler/cml-language/src/main/java/cml/language/features/Model.java
@@ -1,7 +1,5 @@
-package cml.language.features.model;
+package cml.language.features;
 
-import cml.language.features.concept.Concept;
-import cml.language.features.target.Target;
 import cml.language.foundation.elements.ModelElement;
 import cml.language.foundation.elements.ModelElementImpl;
 import cml.language.foundation.elements.Scope;
@@ -43,3 +41,34 @@ public interface Model extends Scope
         return new ModelImpl(modelElement, scope);
     }
 }
+
+class ModelImpl implements Model
+{
+    private final ModelElement modelElement;
+    private final Scope scope;
+
+    ModelImpl(ModelElement modelElement, Scope scope)
+    {
+        this.modelElement = modelElement;
+        this.scope = scope;
+    }
+
+    @Override
+    public Optional<Scope> getParentScope()
+    {
+        return modelElement.getParentScope();
+    }
+
+    @Override
+    public List<ModelElement> getElements()
+    {
+        return scope.getElements();
+    }
+
+    @Override
+    public void addElement(ModelElement element)
+    {
+        scope.addElement(element);
+    }
+}
+
