+++ cml-compiler/cml-language/src/main/java/cml/language/grammar/ModelLoader.java
@@ -2,7 +2,7 @@ package cml.language.grammar;
 
 import cml.io.Console;
 import cml.io.SourceFile;
-import cml.language.features.model.Model;
+import cml.language.features.Model;
 
 import java.util.Optional;
